Refactored to better organize files into shared (among multiple award types) or IDV-specific directories
