+++ src/js/components/awardv2/idv/amounts/AwardsBanner.jsx
