+++ src/js/components/awardv2/idv/amounts/charts/NormalChart.jsx
